LPS-58771 Simpler lifecycle management: avoid the usage of statics and refactor a couple of classes in order to put the behaviour in the same place
