+++ modules/apps/foundation/portal-template/portal-template-soy/src/main/java/com/liferay/portal/template/soy/internal/SoyTemplateResourcesTracker.java
@@ -14,9 +14,20 @@
 
 package com.liferay.portal.template.soy.internal;
 
+import com.liferay.portal.kernel.cache.PortalCache;
+import com.liferay.portal.kernel.cache.SingleVMPoolUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.template.TemplateConstants;
+import com.liferay.portal.kernel.template.TemplateException;
 import com.liferay.portal.kernel.template.TemplateResource;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.template.soy.utils.SoyTemplateResourcesCollector;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -24,45 +35,192 @@ import java.util.concurrent.CopyOnWriteArrayList;
 
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
+import org.osgi.framework.BundleEvent;
 import org.osgi.framework.wiring.BundleCapability;
+import org.osgi.framework.wiring.BundleRevision;
+import org.osgi.framework.wiring.BundleWire;
+import org.osgi.framework.wiring.BundleWiring;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.util.tracker.BundleTracker;
+import org.osgi.util.tracker.BundleTrackerCustomizer;
 
 /**
  * @author Bruno Basto
  */
+@Component(service = SoyTemplateResourcesTracker.class)
 public class SoyTemplateResourcesTracker {
 
-	public static void close() {
-		_bundleTracker.close();
-	}
-
-	public static List<TemplateResource> getAllTemplateResources() {
+	public List<TemplateResource> getAllTemplateResources() {
 		return _templateResources;
 	}
 
-	public static Bundle getBundle(long bundleId) {
+	public Bundle getBundle(long bundleId) {
 		return _bundleMap.get(bundleId);
 	}
 
-	public static Collection<Bundle> getBundles() {
+	public Collection<Bundle> getBundles() {
 		return _bundleMap.values();
 	}
 
-	public static void open(BundleContext bundleContext) {
+	@Activate
+	protected void activate(BundleContext bundleContext) {
 		int stateMask = Bundle.ACTIVE | Bundle.RESOLVED;
 
 		_bundleTracker = new BundleTracker<>(
 			bundleContext, stateMask,
-			new SoyCapabilityBundleTrackerCustomizer(
-				"soy", _bundleMap, _templateResources));
+			new SoyCapabilityBundleTrackerCustomizer("soy"));
 
 		_bundleTracker.open();
 	}
 
+	@Deactivate
+	protected void deactivate() {
+		_bundleTracker.close();
+	}
+
 	private static final Map<Long, Bundle> _bundleMap =
 		new ConcurrentHashMap<>();
-	private static BundleTracker<List<BundleCapability>> _bundleTracker;
 	private static final List<TemplateResource> _templateResources =
 		new CopyOnWriteArrayList<>();
+
+	private BundleTracker<List<BundleCapability>> _bundleTracker;
+
+	private static final class SoyCapabilityBundleTrackerCustomizer
+		implements BundleTrackerCustomizer<List<BundleCapability>> {
+
+		public SoyCapabilityBundleTrackerCustomizer(String namespace) {
+			_soyTofuCacheHandler = new SoyTofuCacheHandler(_portalCache);
+		}
+
+		@Override
+		public List<BundleCapability> addingBundle(
+			Bundle bundle, BundleEvent bundleEvent) {
+
+			_addTemplateResourcesToList(bundle);
+
+			BundleWiring bundleWiring = bundle.adapt(BundleWiring.class);
+
+			List<BundleCapability> bundleCapabilities =
+				bundleWiring.getCapabilities("soy");
+
+			return bundleCapabilities;
+		}
+
+		@Override
+		public void modifiedBundle(
+			Bundle bundle, BundleEvent bundleEvent,
+			List<BundleCapability> bundleCapabilities) {
+
+			removedBundle(bundle, bundleEvent, bundleCapabilities);
+
+			List<BundleCapability> newBundleCapabilities = addingBundle(
+				bundle, bundleEvent);
+
+			bundleCapabilities.clear();
+
+			bundleCapabilities.addAll(newBundleCapabilities);
+		}
+
+		@Override
+		public void removedBundle(
+			Bundle bundle, BundleEvent bundleEvent,
+			List<BundleCapability> bundleCapabilities) {
+
+			List<TemplateResource> removedTemplateResources =
+				_removeBundleTemplateResourcesFromList(bundle);
+
+			_soyTofuCacheHandler.removeIfAny(removedTemplateResources);
+
+			_bundleMap.remove(bundle.getBundleId());
+		}
+
+		private void _addTemplateResourcesToList(Bundle bundle) {
+			BundleWiring bundleWiring = bundle.adapt(BundleWiring.class);
+
+			for (BundleWire bundleWire : bundleWiring.getRequiredWires("soy")) {
+				BundleRevision bundleRevision = bundleWire.getProvider();
+
+				Bundle providerBundle = bundleRevision.getBundle();
+
+				_bundleMap.put(providerBundle.getBundleId(), providerBundle);
+			}
+
+			_bundleMap.put(bundle.getBundleId(), bundle);
+
+			SoyTemplateResourcesCollector soyTemplateResourceCollector =
+				new SoyTemplateResourcesCollector(bundle, StringPool.SLASH);
+
+			try {
+				List<TemplateResource> templateResources =
+					soyTemplateResourceCollector.getTemplateResources();
+
+				templateResources.stream().forEach(
+					templateResource -> {
+						if (!_templateResources.contains(templateResource)) {
+							_templateResources.add(templateResource);
+						}
+					});
+			}
+			catch (TemplateException te) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(
+						"Unable to add template resources for bundle " +
+							bundle.getBundleId(),
+						te);
+				}
+			}
+		}
+
+		private long _getBundleId(String templateId) {
+			int pos = templateId.indexOf(TemplateConstants.BUNDLE_SEPARATOR);
+
+			if (pos == -1) {
+				if (_log.isDebugEnabled()) {
+					String message = String.format(
+						"The templateId \"%s\" does not map to a Soy template",
+						templateId);
+
+					_log.debug(message);
+				}
+
+				return -1;
+			}
+
+			return Long.valueOf(templateId.substring(0, pos));
+		}
+
+		private List<TemplateResource> _removeBundleTemplateResourcesFromList(
+			Bundle bundle) {
+
+			List<TemplateResource> removedTemplateResources = new ArrayList<>();
+
+			Iterator<TemplateResource> iterator = _templateResources.iterator();
+
+			while (iterator.hasNext()) {
+				TemplateResource templateResource = iterator.next();
+
+				long bundleId = _getBundleId(templateResource.getTemplateId());
+
+				if (bundle.getBundleId() == bundleId) {
+					removedTemplateResources.add(templateResource);
+				}
+			}
+
+			_templateResources.removeAll(removedTemplateResources);
+
+			return removedTemplateResources;
+		}
+
+		private static final Log _log = LogFactoryUtil.getLog(
+			SoyCapabilityBundleTrackerCustomizer.class);
+
+		private final PortalCache<HashSet<TemplateResource>, SoyTofuCacheBag>
+			_portalCache = SingleVMPoolUtil.getPortalCache(
+				SoyTemplate.class.getName());
+		private final SoyTofuCacheHandler _soyTofuCacheHandler;
+
+	}
 
 }
